Center contact photo
When the contact photo does not have a quadratic size, the leftmost excerpt is shown. As contact photos usually contains centered pictures of a person's face, it is not helpful to start showing the photo from the left. Instead, it should be centered and the center portion should be shown.