Handle the font size using WebExtension APIs, improve getting of the browser object for stub.xhtml
Font size handling is now setup via the startup of the reducer. Also, reloading the message pane on font size change has been moved to the background script/dedicated api.

The browser object is now obtained in an async fashion, giving us direct access to the APIs we need. We also pass the full context object so that we can use the event based APIs properly.