Improve show involving performance and fix windowId checks.
Loading of show involving was previous done via running two queries then displaying the result. Now we run only the first query and pass the second query to the tab so that it can render the tab and handle the results at the same time. This gives better feedback to the user that something is happening.