Fix: Stop the selected conversation sometimes loosing the expand status when being selected.
Sometimes Thunderbird triggers multiple selection events/load callbacks. As a result we are sometimes getting two Conversation loads and the second one is getting overriden by the first but gaining the fact the conversation has been marked as read. Hence the expanded states are wrong.

This fixes the previous selection recording so that it is saved before the conversation is constructed and rendered, which might take some time.