Move message expand/scroll handling to the reducer, and handle marking as read there as well.
This helps to simplify the code and make it clearer, and add tests. It also moves it out of the message construction which makes that part simpler.

Also moved is mark as read as we need to make sure it happens after the message data is gathered to ensure that read status is not changed beforehand.