Fixes #1913: Show 'reply all' option if reply-to is different from the sender
Mailing lists with 'Mangle From' configuration swaps 'From' with
the mailing list address and keeping the original sender in the
'Reply-To' header.  Currently only 'reply' and 'reply to list'
options are available in this case, but there should be an option
to reply to both the list and the original sender.

Taking the 'reply-to' header into account while checking for multiple
"recipients" to fix the problem.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>